refactor: move logging helpers to internal package
