+++ internal/logging_test.go
@@ -1,4 +1,4 @@
-package dht
+package internal
 
 import (
 	"testing"
